--- src/System.Runtime.Handles/tests/SafeHandle.netstandard.cs
@@ -1,29 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System;
-using Xunit;
-
-public partial class SafeHandle_4000_Tests
-{
-    [Fact]
-    public static void SafeHandle_invalid_close()
-    {
-        MySafeHandle mch = new MySafeHandle();
-        mch.Close();
-        Assert.True(mch.IsClosed);
-        Assert.True(mch.IsInvalid);
-        Assert.False(mch.IsReleased);
-    }
-
-    [Fact]
-    public static void SafeHandle_valid_close()
-    {
-        MySafeHandle mch = new MySafeHandle(new IntPtr(1));
-        mch.Close();
-        Assert.True(mch.IsClosed);
-        Assert.False(mch.IsInvalid);
-        Assert.True(mch.IsReleased);
-    }
-}
